relocate the graph card file into the component folder
